+++ pages/schedule/all.tsx
@@ -1,0 +1,8 @@
+//모든 일정 페이지
+import All from "@schedule/components/All";
+
+const create = () => {
+  return <All />;
+};
+
+export default create;
