--- pages/schedule/create.tsx
@@ -1,8 +1,0 @@
-// 일정 만들기 페이지
-import Create from "@create/components/Create";
-
-const create = () => {
-  return <Create />;
-};
-
-export default create;
