+++ src/schedule/components/ScheduleCard.tsx
