+++ tsconfig.json
@@ -20,7 +20,7 @@
       "@shared/*": ["shared/*"],
       "@login/*": ["login/*"],
       "@findid/*": ["findid/*"],
-      "@create/*": ["create/*"],
+      "@schedule/*": ["schedule/*"],
       "@mypage/*": ["mypage/*"]
     }
   },
